Avoid passing metrics everywhere
