+++ core/lib/zksync_core/src/state_keeper/io/tests/mod.rs
@@ -20,7 +20,6 @@ use crate::{
             create_execution_result, create_transaction, create_updates_manager,
             default_l1_batch_env, default_vm_block_result, Query,
         },
-        types::ExecutionMetricsForCriteria,
         updates::{MiniblockSealCommand, MiniblockUpdates, UpdatesManager},
     },
     utils::testonly::create_l1_batch_metadata,
@@ -392,17 +391,10 @@ async fn test_miniblock_and_l1_batch_processing(
     });
 
     let finished_batch = default_vm_block_result();
-    let batch_tip_metrics = ExecutionMetricsForCriteria::default();
 
     let l1_batch_env = default_l1_batch_env(1, 1, Address::random());
     mempool
-        .seal_l1_batch(
-            None,
-            updates,
-            &l1_batch_env,
-            finished_batch,
-            batch_tip_metrics,
-        )
+        .seal_l1_batch(None, updates, &l1_batch_env, finished_batch)
         .await
         .unwrap();
